+++ Interface/AddOns/MissingTradeSkillsList/Data/Engineering/Engineering_Levels.lua
@@ -39,6 +39,7 @@ MTSL_DATA["Engineering"]["levels"] = {
         ["id"] = 4036,
         ["max_skill"] = 75,
         ["min_xp_level"] = 5,
+        ["rank"] = 1,
     }, -- [1]
     {
         ["trainers"] = {
@@ -66,6 +67,7 @@ MTSL_DATA["Engineering"]["levels"] = {
         ["min_skill"] = 50,
         ["id"] = 4037,
         ["max_skill"] = 150,
+        ["rank"] = 2,
     }, -- [2]
     {
         ["trainers"] = {
@@ -89,6 +91,7 @@ MTSL_DATA["Engineering"]["levels"] = {
         ["min_skill"] = 125,
         ["id"] = 4038,
         ["max_skill"] = 225,
+        ["rank"] = 3,
     }, -- [3]
     {
         ["min_xp_level"] = 35,
@@ -111,5 +114,6 @@ MTSL_DATA["Engineering"]["levels"] = {
         ["min_skill"] = 200,
         ["id"] = 12656,
         ["max_skill"] = 300,
+        ["rank"] = 4,
     }, -- [4]
 }