+++ Interface/AddOns/MissingTradeSkillsList/UI/MainFrames/ToggleButton.lua
@@ -67,7 +67,7 @@ function MTSLUI_TOGGLE_BUTTON:ReanchorToNewParent(parent_frame)
 		gap_top = 0
 	end
 	-- Overwrite parenttframe of Blizzard UI to Skillet-Classic addon if installed and enabled/used
-	if SkilletFrame and SkilletFrame:IsVisible() then
+	if SkilletFrame and SkilletFrame:IsVisible() == true then
 		parent_frame = SkilletFrame
 		gap_left = 0
 		gap_top = 0